+++ libs/web/calendar/ui/quick-info-header/src/lib/quick-info-header.component.ts
@@ -26,6 +26,7 @@ export class QuickInfoHeaderComponent {
   @Input() data!: EjsScheduleModel;
 
   // OUTPUT
+  @Output() remove = new EventEmitter<void>();
   @Output() showEditorDialog = new EventEmitter<void>();
   @Output() closeDialog = new EventEmitter<void>();
 }