+++ libs/web/shared/data-access/models/src/lib/schedule/study-schedule.model.ts
@@ -57,6 +57,7 @@ export class StudyScheduleModel extends ScheduleModel {
       Color: this.color,
       Shift: this.shift,
       FixedSchedules: this.fixedSchedules,
+      ReadOnly: true,
     };
   }
 }